Add an example spec demonstrating basic assertions
Note the use of expect().
All of the chai assertions that we use will be chained from an expect()
call that wraps the object we are testing.

Chai also provides "assert" and "should" (BDD-style) which you may
prefer. For this example, we will use the expect style. You can read
about the other styles on the chai website.